requirements.txt
moviepy
whisper
pydub
pysrt
webvtt-py
spacy
rich
transformers
torch
tqdm
numpy
opencv-python
ffmpeg-python